Add is_win() and is_tie() to tictactoe
Implements is_win() and is_tie() to determine if the game is over

Also, implements player_list including a player with a blank space (" ") for a symbol. This guarantees that the number of players in the player list matches the number of pieces needed to display the board and that the index for the player list matches the index for the pieces list.